Move stats methods to stats.py
